requirements.txt
yfinance>=0.2.50
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
python-dotenv>=1.0.1
loguru>=0.7.2
APScheduler>=3.10.4
pytz>=2024.1
pandas-market-calendars>=4.4.0
pytest>=8.2.0
